requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.2
drf-jwt==1.18.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
